+++ swri_opencv_util/CHANGELOG.rst
@@ -2,6 +2,16 @@
 Changelog for package swri_opencv_util
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.5.0 (2018-10-12)
+------------------
+* Add function for fitting a rotation on 3D point correspondences. (`#524 <https://github.com/swri-robotics/marti_common/issues/524>`_)
+* Contributors: Marc Alban
+
+2.4.0 (2018-10-09)
+------------------
+* Add function to blend two images with alpha channels. (`#522 <https://github.com/swri-robotics/marti_common/issues/522>`_)
+* Contributors: Marc Alban
+
 2.3.0 (2018-05-25)
 ------------------